+++ AGENTS-tool-adapters.md
@@ -74,7 +74,7 @@ eager-load), `lib/axn/tools/adapter_roots.rb`, `lib/axn/core/tool_declaration.rb
   with one documented **looser** case (an invalid literal `default:`). Surface the caveat; don't fight it.
   A deep subfield under a `model:`/non-object parent is omitted with a `logger.warn` — pass it through.
 
-Source: `lib/axn/core/schema_reflection.rb`, `lib/axn/reflection/schema.rb`.
+Source: `lib/axn/core/schema_reflection.rb`, `lib/axn/internal/reflection/schema.rb`.
 
 ## Value serialization
 
@@ -104,7 +104,7 @@ Source: `lib/axn/core/schema_reflection.rb`, `lib/axn/reflection/schema.rb`.
 - Keep the two guarantees apart: encodability is unconditional, declared-shape is what the flag buys.
   `reject_opaque: false` never means "might not be JSON" — that is why it isn't named `strict:`.
 
-Source: `lib/axn/extensions/serialization.rb` (the renderer itself is `lib/axn/reflection/values.rb`, core-internal).
+Source: `lib/axn/extensions/serialization.rb` (the renderer itself is `lib/axn/internal/reflection/values.rb`, core-internal).
 
 ## Per-adapter configuration
 
@@ -218,7 +218,7 @@ Core source entry points (resolve with `bundle show axn`):
 - `lib/axn/extensions.rb` — `Axn::Extensions.best_effort`, `.config`, `.owned_failure?` (extension-author surface).
 - `lib/axn/tools/registry.rb`, `lib/axn/tools/adapter_roots.rb`, `lib/axn/core/tool_declaration.rb` —
   membership, `tool_name`.
-- `lib/axn/core/schema_reflection.rb`, `lib/axn/reflection/schema.rb`, `lib/axn/reflection/values.rb` — reflection.
+- `lib/axn/core/schema_reflection.rb`, `lib/axn/internal/reflection/schema.rb`, `lib/axn/internal/reflection/values.rb` — reflection.
 - `lib/axn/configurable.rb` — `config_namespace`, `resolve_override_for`, `overrides`.
 - `lib/axn/tools/invoker.rb` — the tool call path.
 - `lib/axn/core/ambient_context.rb` — ambient filtering/resolution.
