+++ CHANGELOG.md
@@ -105,7 +105,7 @@
 * [FEAT] `tag` (high-cardinality) and `dimension` (bounded) declare per-action observability facets, resolved from inputs before the body (`from: :inputs`) or from the settled result (`from: :result`). Each facet flows to the `axn.call` OpenTelemetry span, the notification payload, `emit_metrics`, `auto_log` output (as SemanticLogger tags, or a readable suffix under any other logger), the `on_exception` report (`context[:tags]`/`context[:dimensions]`, now reserved keys), and Sidekiq per-job tags (selected by `Axn.config.sidekiq_job_tag_sources`, per-class overridable).
 * [BREAKING] `on_success` now fires after the enclosing transaction commits (immediately when none is open) and is skipped on rollback, so an action nested inside another's transaction no longer fires its side effect when the outer transaction later rolls back. Implemented via `ActiveRecord.after_all_transactions_commit` (requires ActiveRecord 7.2+; inline without ActiveRecord). Failure-path callbacks are unaffected.
 * [FEAT] `on_enqueue_all` is a once-per-run callback for `enqueues_each`/`enqueue_all` fan-out, firing after the loop completes with the enqueued `count:` and a `sources:` hash (flexible arity). It's error-isolated, so a raising callback can't abort the fan-out.
-* [BUGFIX] Files that reference another file's constant at RUNTIME now require it, instead of relying on the top-level `axn` entrypoint's require order. Loading the adapter-facing `axn/reflection` on its own and serializing any Hash or Array raised `NameError`; the same latent gap existed in the call logger, exception-report formatting, facet coercion, async serialization, and `FormObject` (which uses `ActiveModel`, a hard gem dependency, without declaring it). Also in the property-name rules, the schema builder, the shape-graph reader and the field-name conventions, where a process that had loaded only `axn/reflection/property_names` (or `axn/extensions/serialization`) raised `NameError` on its first validated projection — the ownership reader those rules gate names on, and the `model:` id convention both the rules and the builder key off. The contract is now derived rather than listed: `spec/axn/standalone_require_spec.rb` reads every constant each file's code references out of its own parse tree, so a reference added later is covered automatically, and the references that legitimately point UP (a lower file calling back into the layer that composes it; reflection reaching for something only a declared axn class has) are enumerated with reasons and may only shrink. The hand-written list it replaces was stale within days of being written — it named four constants for a file that had grown two more, and the gap it existed to catch went unnoticed.
+* [BUGFIX] Files that reference another file's constant at RUNTIME now require it, instead of relying on the top-level `axn` entrypoint's require order. Loading the adapter-facing `axn/internal/reflection` on its own and serializing any Hash or Array raised `NameError`; the same latent gap existed in the call logger, exception-report formatting, facet coercion, async serialization, and `FormObject` (which uses `ActiveModel`, a hard gem dependency, without declaring it). Also in the property-name rules, the schema builder, the shape-graph reader and the field-name conventions, where a process that had loaded only `axn/internal/reflection/property_names` (or `axn/extensions/serialization`) raised `NameError` on its first validated projection — the ownership reader those rules gate names on, and the `model:` id convention both the rules and the builder key off. The contract is now derived rather than listed: `spec/axn/standalone_require_spec.rb` reads every constant each file's code references out of its own parse tree, so a reference added later is covered automatically, and the references that legitimately point UP (a lower file calling back into the layer that composes it; reflection reaching for something only a declared axn class has) are enumerated with reasons and may only shrink. The hand-written list it replaces was stale within days of being written — it named four constants for a file that had grown two more, and the gap it existed to catch went unnoticed.
 * [BUGFIX] Rendering a result through `Axn::Extensions::Serialization.render` now raises the new `Axn::Extensions::Serialization::UnserializableValue` for a self-referential exposed value instead of recursing to `SystemStackError`. A cycle has no JSON representation, so this is a serialization failure rather than something to render — and being an `ArgumentError`, a serving adapter's existing `rescue StandardError` maps it to an error response, which `SystemStackError` (outside `StandardError`) escaped entirely. The message names the path to the offending value (`items[1]`, `node.child`), and the guard tracks ancestry, so the same object referenced twice as siblings (a diamond) is not a false positive. Also covers an object whose own `as_json`/`to_h` projection points back at it, which builds a fresh Hash per call and so could not be caught by tracking the produced Hash.
 * [BUGFIX] A self-referential (cyclic) Array or Hash no longer crashes `.call` with `SystemStackError`. Every walker that formats caller-supplied values for observability — the call logger, exception-report context, facet coercion, and sensitive-member masking — now cycle-guards its recursion the way Ruby's own `#inspect` does, rendering `[...]`/`{...}` for a container already open on the path (a container merely repeated among siblings still renders in full). Since `SystemStackError` isn't a `StandardError`, it previously bypassed `fail!`/`fails_on`/`on_exception` and escaped to the caller. `ActiveSupport::ParameterFilter` (which any `sensitive:` declaration puts in the path, and which axn cannot guard from inside) is handled by retrying the slice on a decycled copy — so a cyclic field no longer costs the whole log line and the whole `on_exception` report context, and acyclic data pays nothing for the fallback.
 * [BUGFIX] An `ActionController::Parameters` input holding a self-referential value no longer costs the whole `on_exception` report (or the whole log line). `Parameters#to_unsafe_h` recursively rebuilds every nested container, so it blows the stack before axn's cycle guard can observe the repeated container; the conversion is now attempted and falls back to a `{...}` placeholder for that one value. Reachable only by mutating an already-nested Array in place, since `Parameters.new`/`[]=` convert eagerly and raise at assignment.
