+++ lib/axn/internal/reflection/schema.rb
@@ -3,9 +3,9 @@
 require "date"
 require "time"
 
-require "axn/reflection/subfield_tree"
+require "axn/internal/reflection/subfield_tree"
 # A property name in an emitted schema is the canonicalization's answer, so the builder cannot load without it.
-require "axn/reflection/values"
+require "axn/internal/reflection/values"
 
 # The `model:` id convention and the conditional-gate keys are both read on the build path, so the builder
 # cannot load without their owner either.
