+++ lib/axn/internal/reflection/subfield_contradictions.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require "axn/reflection/subfield_tree"
-require "axn/reflection/schema"
+require "axn/internal/reflection/subfield_tree"
+require "axn/internal/reflection/schema"
 
 module Axn
   module Reflection
