+++ spec/axn/internal/reflection/subfield_contradictions_spec.rb
